Remove support for configuring 'package'
Managing the package parameter is more trouble than it is worth, as there is no nice way to programmatically reset it back to default/empty. Assigning $null or an empty string as the value for Package breaks the firewall rule, even though when viewing via the GUI it appears to work (i.e. correctly reselects the "Apply to all programs and application packages" option). 

This is not likely functionality we will need anyway.