Add function for converting subnet mask to prefix length
Add a function that converts a subnet mask (255.255.255.0) to a prefix length (24) when it is returned to Puppet.

Powershell always returns IPv4 network addresses with a /255.255.255.0 style network mask value, even though it will let you set it using prefix length (/24, for example).

Puppet sees this discrepancy as a change to be made (because we define 24 in Puppet but Powershell returns 255.255.255.0 when getting resource information). The 255.255.255.0 representation should be converted before it is returned to Puppet, so Puppet correctly observes no change is to be made.